+++ src/main/java/outfoot/outfootserver/confirm/service/ConfirmService.java
@@ -14,6 +14,9 @@ import outfoot.outfootserver.confirm.exception.ConfirmErrorCode;
 import outfoot.outfootserver.confirm.exception.ConfirmException;
 import outfoot.outfootserver.confirm.repository.ConfirmRepository;
 import outfoot.outfootserver.files.TestFileUploader;
+import outfoot.outfootserver.member.domain.Member;
+import outfoot.outfootserver.member.exception.AuthErrorCode;
+import outfoot.outfootserver.member.exception.AuthException;
 
 @Service
 @RequiredArgsConstructor
@@ -24,7 +27,7 @@ public class ConfirmService {
     private final TestFileUploader fileUploader;
     private final String path = "confirm/";
     @Transactional
-    public ConfirmResponse saveConfirm(CheckPage checkPage, ConfirmRequest dto) {
+    public ConfirmResponse saveConfirm(CheckPage checkPage, ConfirmRequest dto, Member member) {
 
 //        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 //        LocalDateTime now = LocalDateTime.now();
@@ -43,7 +46,7 @@ public class ConfirmService {
             imageUrl = fileUploader.uploadFile(image, path);
         }
 
-        Confirm confirm = confirmRepository.save(ConfirmRequest.toConfirm(dto, checkPage, imageUrl));
+        Confirm confirm = confirmRepository.save(ConfirmRequest.toConfirm(dto, checkPage, imageUrl, member));
 
         long likeCount = confirm.getLikeCount();
         long dislikeCount = confirm.getDisLikeCount();
@@ -52,8 +55,12 @@ public class ConfirmService {
     }
 
     @Transactional
-    public ConfirmUpdateResponse updateConfirm(Long id, UpdateConfirmRequest dto) {
-        Confirm confirm = findById(id);
+    public ConfirmUpdateResponse updateConfirm(Long confirmId, UpdateConfirmRequest dto, Member member) {
+        Confirm confirm = findById(confirmId);
+
+        if (confirm.getMember() != member) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
+        }
 
         String imageUrl = null;
         if (dto.image() != null && !dto.image().isEmpty()){
@@ -70,17 +77,24 @@ public class ConfirmService {
     }
 
     @Transactional
-    public void deleteConfirm(Long confirmId) {
+    public void deleteConfirm(Long confirmId, Member member) {
         Confirm confirm = findById(confirmId);
+
+        if (confirm.getMember() != member) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
+        }
 
         if (confirm.getImageUrl() != null) {
             fileUploader.deleteFile(confirm.getImageUrl(), "confirm");
+        }
+        if (confirm.getMember() != member) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
         }
         confirmRepository.delete(confirm);
     }
 
 
-    public ConfirmResponse findConfirm(Long id){
+    public ConfirmResponse findConfirm(Long id, Member member){
         Confirm confirm = findById(id);
         long likeCount = confirm.getLikeCount();
         long dislikeCount = confirm.getDisLikeCount();
@@ -91,5 +105,6 @@ public class ConfirmService {
     public Confirm findById (Long confirmId) {
         return confirmRepository.findById(confirmId)
                 .orElseThrow(() -> new ConfirmException(ConfirmErrorCode.CONFIRM_NOT_FOUND));
+
     }
 }
